EXtrapolation DETection tool

Assesses univariate (Type I) and combinatorial (Type II) extrapolation between a reference system (ref) and a projection system (p). See Mesgaran et al. (2014) for an explanation. This function is an updated version of some original code from the ecospat package (Broennimann et al. 2016).

ExDet(ref, tg, xp)

Arguments

ref
Reference data. A data.frame with the values of the variables (i.e. columns) for each sample unit (e.g. segments).

tg
Target data. A data.frame with the values of the variables (i.e. columns) for each point of the prediction extent.

xp
Character string. Names of the covariates of interest.

Value

Returns a tibble with four columns. (1) ExDet: Value of the ExDet metric (negative for univariate extrapolation, >1 for combinatorial extrapolation, within the range 0-1 for analogue conditions). (2) mic_univariate: Integer indicating the covariate with the largest contribution to univariate extrapolation (most influential covariate, MIC). (3) mic_combinatorial: Integer indicating the covariate with the largest contribution to combinatorial extrapolation. (4) mic: most influential covariate.

References

Mesgaran, M.B., Cousens, R.D. & Webber, B.L. (2014) Here be dragons: a tool for quantifying novelty due to covariate range and correlation change when projecting species distribution models. Diversity & Distributions, 20: 1147-1159, DOI: 10.1111/ddi.12209

Broennimann O, Di Cola V, Guisan A (2016). ecospat: Spatial Ecology Miscellaneous Methods. R package version 2.1.1. https://CRAN.R-project.org/package=ecospat.