Summary of extrapolation

Displays a tabular summary of the geographic extent of extrapolation. This is calculated as the number (and proportion) of prediction locations (i.e. grid cells) subject to extrapolation.

summarise_extrapolation(extrapolation.object, covariate.names = NULL,
  extrapolation = TRUE, mic = TRUE, print.precision = 2)

Arguments

extrapolation.object
Output object from a run of compute_extrapolation.

covariate.names
Character string. Names of the covariates of interest.

extrapolation
Logical. Whether to return a summary of univariate/combinatorial extrapolation. Defaults to TRUE.

mic
Logical. Whether to return a summary of the most influential covariates (MIC) - see compute_extrapolation. Defaults to TRUE.

print.precision
Integer. Number of significant figures to be used when printing the summary. Default value of 2.

Value

Prints a summary table in the R console. In addition, if assigned to an object, returns a list with the table values (.n = number of locations, .p = corresponding percentage).

References

Bouchet PJ, Miller DL, Roberts JJ, Mannocci L, Harris CM and Thomas L (2019). From here and now to there and then: Practical recommendations for extrapolating cetacean density surface models to novel conditions. CREEM Technical Report 2019-01, 59 p. https://research-repository.st-andrews.ac.uk/handle/10023/18509

Mesgaran MB, Cousens RD, Webber BL (2014). Here be dragons: a tool for quantifying novelty due to covariate range and correlation change when projecting species distribution models. Diversity & Distributions, 20: 1147-1159, DOI: 10.1111/ddi.12209

See also

Examples

library(dsmextra)

# Load the Mid-Atlantic sperm whale data (see ?spermwhales)
data(spermwhales)

# Extract the data
segs <- spermwhales$segs
predgrid <- spermwhales$predgrid

# Define relevant coordinate system
my_crs <- sp::CRS("+proj=aea +lat_1=38 +lat_2=30 +lat_0=34 +lon_0=-73 +x_0=0
 +y_0=0 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0")

# Assess extrapolation in the multivariate space defined by five covariates
spermw.extrapolation <- compute_extrapolation(segments = segs,
      covariate.names = c("Depth", "DistToCAS", "SST", "EKE", "NPP"),
      prediction.grid = predgrid,
      coordinate.system = my_crs,
      print.summary = FALSE,
      save.summary = TRUE,
      print.precision = 2)

#> Computing ...

#> Done!

# Summarise extrapolation
spermw.summary <- summarise_extrapolation(extrapolation.object = spermw.extrapolation,
     covariate.names = c("Depth", "DistToCAS", "SST", "EKE", "NPP"),
     extrapolation = TRUE,
     mic = TRUE,
     print.precision = 2)

#>
#>
#> Table: Extrapolation
#>
#> --------------  ------------  ------------
#> Analogue        n = 5,175     97.92 %
#> -----------     -----------   -----------
#> Univariate      n = 65        1.23 %
#> Combinatorial   n = 45        0.85 %
#> -----------     -----------   -----------
#> Sub-total       n = 110       2.08 %
#> -----------     -----------   -----------
#> Total           n = 5,285     100 %
#> --------------  ------------  ------------
#>
#>
#> Table: Most influential covariates (MIC)
#>
#> --------------  ------------  ------------  ------------
#> Univariate      EKE           n = 24        0.45 %
#> Univariate      SST           n = 17        0.32 %
#> Univariate      Depth         n = 16        0.3 %
#> Univariate      NPP           n =  6        0.11 %
#> Univariate      DistToCAS     n =  2        0.04 %
#> -----------     -----------   -----------   -----------
#> Sub-total                     n = 65        1.23 %
#> -----------     -----------   -----------   -----------
#> Combinatorial   Depth         n = 39        0.74 %
#> Combinatorial   DistToCAS     n =  6        0.11 %
#> -----------     -----------   -----------   -----------
#> Sub-total                     n = 45        0.85 %
#> -----------     -----------   -----------   -----------
#> Total                         n = 110       2.08 %
#> --------------  ------------  ------------  ------------

print(spermw.summary)

#> $extrapolation
#> $extrapolation$univariate.n
#> [1] 65
#>
#> $extrapolation$univariate.p
#> [1] 1.229896
#>
#> $extrapolation$combinatorial.n
#> [1] 45
#>
#> $extrapolation$combinatorial.p
#> [1] 0.8514664
#>
#> $extrapolation$analogue.n
#> [1] 5175
#>
#> $extrapolation$analogue.p
#> [1] 97.91864
#>
#>
#> $mic
#> $mic$Depth
#> $mic$Depth$.n
#> [1] 16
#>
#> $mic$Depth$.p
#> [1] 0.3027436
#>
#>
#> $mic$DistToCAS
#> $mic$DistToCAS$.n
#> [1] 2
#>
#> $mic$DistToCAS$.p
#> [1] 0.03784295
#>
#>
#> $mic$EKE
#> $mic$EKE$.n
#> [1] 24
#>
#> $mic$EKE$.p
#> [1] 0.4541154
#>
#>
#> $mic$NPP
#> $mic$NPP$.n
#> [1] 6
#>
#> $mic$NPP$.p
#> [1] 0.1135289
#>
#>
#> $mic$SST
#> $mic$SST$.n
#> [1] 17
#>
#> $mic$SST$.p
#> [1] 0.3216651
#>
#>
#> $mic$Depth
#> $mic$Depth$.n
#> [1] 39
#>
#> $mic$Depth$.p
#> [1] 0.7379376
#>
#>
#> $mic$DistToCAS
#> $mic$DistToCAS$.n
#> [1] 6
#>
#> $mic$DistToCAS$.p
#> [1] 0.1135289
#>
#>
#>